Example of drawing a custom sprite HUD overlay
This is an example of an entity which will render a sprite (possibly animated)
to the HUD. It's used in 'The Core' for chapter titles. You can see a demo
video here: https://cdn.discordapp.com/attachments/529986135896883211/756774334822875136/Sprite_overlay_demo.mp4

This entity will not save any data, but it prevents re-displaying itself when loading a save
by removing itself after firing.

If you wanted to make this a toggle-able overlay instead of a temporary time-based one,
you would need to add Save/Restore to the entity and a flag to show if the entity is "on"
or not. Depending on your implementation, you would also need to fire the message as soon
as the entity spawns, if it's saved in the "on" state (or has a "start on" spawnflag).
Carrying the overlay flag over across multiple maps would be more difficult, I'm not sure what
would be involved in doing that. The easiest option would be to put that burden onto the mapper
to use globals, trigger_auto, and/or a "start on" spawnflag on the entity to make sure the HUD
icon is persisted across maps.

Here's the FGD definition for this entity, if you would like to use it:

@PointClass base(Targetname, Target) iconsprite("sprites/env.spr") = env_sprite_overlay : "Sprite Overlay"
[
    framerate(string) : "Framerate" : "15"
    model(sprite) : "Sprite Name" : ""
    time(integer) : "Display seconds (-1 = one animation)": -1
    halign(integer) : "Horizontal alignment (%)" : 50
    valign(integer) : "Vertical alignment (%)" : 50
]