Add save/restore to game_player_equip
This simple change shows an example of adding save/restore to any entity.

game_player_equip does not persist across save games in the current SDK,
so the weapon names/values are lost if the player dies or loads a save.
This makes it only useful for trigger_auto scenarios or cases where it's
extremely unlikely for the player to save before the player equip is triggered.
Adding save/restore to the entity fixes this issue.

Adding save data to an entity is quite easy, as you can see in this change set.
You define the `Save` and `Restore` methods as virtual, and add the
`m_SaveData` array. You set up that array, and then add a call to the
`IMPLEMENT_SAVERESTORE` macro with the class name and the base class name.
It's as easy as that!

When setting up the `m_SaveData` array, you have two options:

For a value type (e.g. float, int, string, etc):
    DEFINE_FIELD( <class name>, <field name>, <field type> ),

For an array type (e.g. float[], int[] string[], etc):
    DEFINE_ARRAY( <class name>, <field name>, <field type>, <array length> ),
As the type description is static, the array length must be fixed at compile time.

It's also possible (but not too common) to implement your own custom Save/Restore
functions instead of calling `IMPLEMENT_SAVERESTORE`. This might be useful if you
want to save something that the default implementation cannot support (perhaps a
linked list or vector type, or something more complicated), or want to run some
custom logic after performing the restore. For more detail about custom Save/Restore
in the SDK, see:

- The `IMPLEMENT_SAVERESTORE` macro, in saverestore.h
- `Save` and `Restore` for `CBaseMonster`, in monsters.cpp
- `Save` and `Restore` for `CBasePlayer`, in player.cpp
- `Save` and `Restore` for `CGlobalState`, in world.cpp