Adding a new sprite resolution to the HUD
This example shows how to add a custom sprite resolution to
the HUD. Half-Life ships with two HUD resolutions: 320, and 640.
This might have been reasonable in 1998, but these days you can't
even select a resolution lower than 640x480 in the Steam version
of the game, so the 640 HUD might as well be the only resolution
that exists. I think the size of it is fine for 1080p (except for
the size of the revolver crosshair), but it's not unusual to see
somebody request HUD sizes for larger resolutions. It's probably
way too small at 4k, though I've never tried that myself.

The HUD is based on sprites of a fixed size, so scaling is not
possible without much larger changes to the code. But we can extend
the current logic to add additional resolutions without much hassle.

The following change adds a HUD that appears on resolutions of 1920
or larger. Because this is just an example, all I did was double the
size of the HUD sprites using an image upscaler. They don't look great,
but whatever. The train/pain/logo sprites aren't even upscaled, because
animated sprite palettes are hard and I don't want to deal with that.

The logic can be extended to resolutions of any size, and you could even
have one for every resolution you could think of. Editing all the HUD
sprites is very, very tedious though, so it's probably a huge waste of
time. To properly set a resolution up, you need to do the following:

- Create HUD sprites at your desired size (they are stored in the
  hud 1-9 sprites, but you can split them up into more files if you need to)
- Set up your new resolution in hud.txt - you must add a line for every HUD
  sprite in the expected format, and increase the number at the top of the
  file by the number of sprites you added. hud.txt lines are formatted like:
  `<sprite name> <resolution> <file name> <x> <y> <width> <height>`
- Modify EVERY weapon_*.txt file in much the same way as hud.txt.
  This takes ages.
- Create the following 3 sprites, with X replaced with your resolution. These
  sprites are animated, you need to make sure each frame in your sprite
  matches each frame in the originals:
  - X_logo.spr (only used in demo mode, just copy the existing one)
  - X_pain.spr (displayed when the player is hurt)
  - X_train.spr (displayed when the player is using trains)

You can see an example of this demo in action here:
https://cdn.discordapp.com/attachments/529986135896883211/757221507830251530/Double_size_1920_hud.png
The sprites are bad because it's just a dumb upscale, but notice that the
HUD images are twice as big as they would normally be in 1920x1080 resolution.

If you want to compile this code yourself to try it, you will
need to put the files from the upscaled sprite pack (see included 7z file)
in your mod directory. It contains all the upscaled sprites with the
right file names, and the hud.txt/weapon_*.txt files with the appropriate
edits already done.