Add the Dark Souls YOU DIED message on death
OK, this one's a bit of a joke, but whatever.
I was going through the HUD classes and found the
"death notice". This is a multiplayer only HUD
element that prints the "kill log" in deathmatch.

This change only works properly in single player,
and the death notice HUD element has been replaced
by Dark Souls' iconic YOU DIED message. It might
be a little buggy, but honestly, who cares?

The HUD code is very similar to the code in the
env_sprite_overlay branch, see that commit for
a better explanation of what's going on here.

Here's a video of it in action:
https://cdn.discordapp.com/attachments/291678871856742400/757911810618818620/Half_Life_2020.09.22_-_20.23.45.16_2.mp4

If you want to use this you might want the sound
and sprite file. I'm not putting them into GitHub
for copyright reasons, but hopefully you can download
them from here:

https://cdn.discordapp.com/attachments/529986135896883211/757950469216927877/you-died-resources.7z

(If this link is down, just find the black background
YOU DIED video on YouTube and make an animated sprite out
of it with the ADDITIVE setting. Follow The303's tutorial
on animated sprites: http://the303.org/tutorials/gold_sprite_p5.htm )

You could probably update this to work on multiplayer
as well, you would just need to properly filter so that
the message only appears when the client's player has
died, rather than ANY player.